requirements.txt
absl-py==2.3.1
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.60.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
jax==0.7.2
jaxlib==0.7.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mediapipe==0.10.14
ml_dtypes==0.5.3
narwhals==2.5.0
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pip==24.0
protobuf==4.25.8
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scipy==1.16.2
setuptools==65.5.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.2
streamlit==1.49.1
streamlit-image-comparison==0.0.4
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
